requirements.txt
google-generativeai
pillow
pandas
openpyxl
python-dotenv
numpy
rapidfuzz
pdf2image
pypdf
PyMuPDF
requests
tqdm
pytesseract
